Add files via upload
night time traffic surveillance video as its input. The system detects pairs and tracks the headlights in  the  video.  The headlight  detection  is  done by  extracting two features namely  reflection intensity map and reflection suppressed map based on the analysis of light attenuation model. This feature extraction is mainly done for discriminating the headlight from the reflection on the road surface. The headlight is tracked and paired using yet effective reasoning algorithm. Experimental results show that the proposed method can robustly detect and pair the vehicle headlight in night scenes. The speed of the vehicles can also be calibrated.